libtests: Address -Wsign-compare warnings
Most of these warnings were between int or ssize_t and size_t.
In particular, read() and write() return ssize_t and comparing
that value with the sizeof a buffer is a source of these warnings.

With the type change, some printf() specifiers needed to change also.

+++ testsuites/libtests/block06/init.c
@@ -395,7 +395,7 @@ static bool
 bdbuf_disk_ioctl_process (bdbuf_disk* bdd, rtems_blkdev_request* req)
 {
   bool result = true;
-  int  b;
+  uint32_t  b;
 
   /*
    * Perform the requested action.
@@ -1725,7 +1725,7 @@ bdbuf_tester (void)
 {
   bdbuf_task_control        tasks[BDBUF_TEST_TASKS];
   rtems_task_priority       old_priority;
-  int                       t;
+  size_t                    t;
   bool                      passed = true;
   rtems_status_code         sc;
 
@@ -1773,9 +1773,9 @@ bdbuf_tester (void)
    */
   for (t = 0; (t < BDBUF_TEST_NUM) && passed; t++)
   {
-    printf ("test %d: %s\n", t + 1, bdbuf_tests[t].label);
+    printf ("test %zu: %s\n", t + 1, bdbuf_tests[t].label);
     passed = bdbuf_tests[t].test (tasks);
-    printf ("test %d: %s\n", t + 1, passed ? "passed" : "failed");
+    printf ("test %zu: %s\n", t + 1, passed ? "passed" : "failed");
   }
 }
