+++ testsuites/libtests/termios04/init.c
@@ -55,8 +55,8 @@ void write_helper(
   const char *c
 )
 {
-  size_t   len;
-  int      rc;
+  ssize_t   len;
+  ssize_t  rc;
   
   len = strlen( c );
   printf( "Writing: %s", c );
